+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - new function to check remote console version to enable/disable functionality or feature of the cli
+- support to import service account from raw file if supported by the remote server
 
 ## [v0.17.2] - 2025-03-04